Refactor cluster routing logic into its own file
Pulls all the logic related to determining and tracking which
node(s) should handle a request into a separate file.

Additional changes:
* Remove redundant check_connection function.
